Remove DBSCAN and consolidate nodes
